+++ app/src/main/java/com/carvajal/lucas/supertaster/data/RecipeImage.kt
@@ -11,6 +11,6 @@ import androidx.room.ForeignKey
     ]
 )
 data class RecipeImage(
-    @ColumnInfo(name = "recipe_id") val recipeId: Int,
+    @ColumnInfo(name = "recipe_id") val recipeId: Long,
     @ColumnInfo(name = "location") val location: String,
 )
